% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catt_gt_discrete.R
\name{catt_gt_discrete}
\alias{catt_gt_discrete}
\title{Group-Time Conditional Average Treatment Effects Given a Discrete Covariate}
\usage{
catt_gt_discrete(
  yname,
  tname,
  idname,
  gname,
  zname,
  xformla,
  data,
  zeval,
  gteval = NULL,
  control_group = "nevertreated",
  anticipation = 0,
  alp = 0.05,
  biters = 1000,
  uniformall = FALSE,
  cores = 1
)
}
\arguments{
\item{yname}{The name of the outcome}

\item{tname}{The name of the time periods}

\item{idname}{The name of the cross-sectional IDs}

\item{gname}{The name of the groups.
"G = 0" indicates the never treated group.}

\item{zname}{The name of the scalar discrete covariate
for which the group-time conditional average treatment effects are estimated}

\item{xformla}{A formula for the covariates to include in the model.
It should be of the form `~ X1 + X2`.
`xformla` should include `zname` as a covariate.}

\item{data}{The name of data.frame that contains the balanced panel data}

\item{zeval}{The vector of the evaluation points z}

\item{gteval}{The vector or matrix of the evaluation points g and t.
If it is a vector, the first and second elements indicate g and t, respectively.
If it is a matrix, the first and second columns indicate g's and t's, respectively.
Default is `NULL`, and `gteval` is automatically constructed.}

\item{control_group}{Which units to use the control group.
Options are "nevertreated" and "notyettreated".
Default is "nevertreated".}

\item{anticipation}{The number of time periods before participating in the
treatment where units can anticipate participating in the treatment and
therefore it can affect their untreated potential outcomes.
Default is 0.}

\item{alp}{The significance level. Default is 0.05.}

\item{biters}{The number of bootstrap iterations to use.
Default is 1000.}

\item{uniformall}{Boolean for whether or not to perform the uniform inference over (g, t, z).
Default is `FALSE`, and the uniform inference only over z is performed.}

\item{cores}{The number of cores to use for parallel processing.
The number of available cores can be checked with parallel::detectCores().
Default is 1.}
}
\value{
A list that contains the following elements:
\item{Estimate}{A data.frame that contains the following elements: \cr
g: The group. \cr
t: The period. \cr
z: The covariate value. \cr
est: The doubly robust estimate of CATT. \cr
se: The standard error. \cr
ci_lower: The lower bound of the UCB via multiplier bootstrapping. \cr
ci_upper: The upper bound of the UCB via multiplier bootstrapping.}
\item{Figure}{A list that contains the ggplot elements for the UCB}
}
\description{
`catt_gt_discrete` computes doubly robust uniform confidence bands for
the group-time conditional average treatment (CATT) function given a discrete
pre-treatment covariate in the staggered difference-in-differences setup of
Callaway and Sant'Anna (2021).
See Imai, Qin, and Yanagi (2023) for details.
}
\examples{
\dontrun{
set.seed(1)
data <- datageneration(n = 1000, tau = 4, continuous = FALSE)
est <- catt_gt_discrete(yname = "Y",
                        tname = "period",
                        idname = "id",
                        gname = "G",
                        zname = "Z",
                        xformla = ~ Z,
                        data = data,
                        zeval = c(-1, 0, 1),
                        gteval = c(2, 2),
                        control_group = "nevertreated",
                        anticipation = 0,
                        alp = 0.05,
                        biters = 1000,
                        uniformall = FALSE,
                        cores = 1)
}

}
\references{
Callaway, B., & Sant’Anna, P. H. (2021).
Difference-in-differences with multiple time periods.
Journal of Econometrics, 225(2), 200-230.

Imai, S., Qin, L., & Yanagi, T. (2023).
Doubly Robust Uniform Confidence Bands
for Group-Time Conditional Average Treatment Effects
in Difference-in-Differences.
arXiv preprint arXiv:2305.02185.
}
