% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datageneration.R
\name{datageneration}
\alias{datageneration}
\title{Data generating function}
\usage{
datageneration(n, tau, continuous = TRUE)
}
\arguments{
\item{n}{The number of cross-sectional units}

\item{tau}{The length of time series such that tau > 2}

\item{continuous}{Boolean for whether or not to consider a continuous covariate.
If it is TRUE (resp. FALSE), a continuous (resp. discrete) covariate is generated.
Default is TRUE.}
}
\value{
A data.frame that contains the following elements.
\item{id}{The unit index}
\item{period}{The time period index}
\item{Y}{The outcome}
\item{G}{The group}
\item{Z}{The scalar continuous covariate}
}
\description{
Data generating function
}
\examples{
set.seed(1)
data <- datageneration(n = 1000, tau = 4, continuous = TRUE)

}
